+++ doc/changelog.md
@@ -9,12 +9,12 @@ in the next tagged version.  Get them with
 ```sh
 go install github.com/magisterquis/curlrevshell@dev
 ```
-- `-version`: Took a first whack at version info.
 - `-h`: Invisible changes to print help a bit more nicely in strange conditions.
 - New and improved tests which now actually run curlrevshell.
 - Updated dependencies.
 - [`-tls-certificate-cache`](./flags.md#-tls-certificate-cache): Added a
   "simple" command to generate a new cert and key.
+- New friendly welcome message with the current version.
 
 
 `v0.0.1-beta.7` (2024-10-22)